refactor(middleware): restructure folders to group external APIs
