+++ middleware/stemmaweb_middleware/controller/auth/service.py
@@ -3,7 +3,7 @@ from authlib.integrations.flask_client import OAuth
 
 import stemmaweb_middleware.constants as constants
 from stemmaweb_middleware.models import StemmawebUser
-from stemmaweb_middleware.stemmarest import APIClient
+from stemmaweb_middleware.resources.base import APIClient
 
 from . import models
