+++ middleware/stemmaweb_middleware/resources/base/__init__.py
@@ -1,0 +1,3 @@
+from .api_client import APIClient
+
+__all__ = ["APIClient"]
