+++ middleware/stemmaweb_middleware/resources/stemmarest/__init__.py
@@ -1,0 +1,3 @@
+from stemmaweb_middleware.resources.base.api_client import APIClient
+
+__all__ = ["APIClient"]
