+++ middleware/stemmaweb_middleware/resources/stemmarest/permissions/predicates.py
