--- middleware/stemmaweb_middleware/stemmarest/__init__.py
@@ -1,3 +1,0 @@
-from .stemmarest_client import APIClient
-
-__all__ = ["APIClient"]
